First steps towards runmanager refactoring.
Separating functionality into different files as much as convenient.
