+++ runmanager/__main__.py
@@ -61,6 +61,10 @@ from labscript_utils import dedent
 from zprocess import raise_exception_in_thread
 import runmanager
 import runmanager.remote
+import runmanager.compiler as compiler
+import runmanager.group_manager as group_manager
+from runmanager.exceptions import ExpansionError
+from runmanager.widgets import RunmanagerColors, FingerTabWidget, TableView, TreeView, TabToolButton, AlternatingColorModel
 
 from qtutils import (
     inmain,
@@ -102,18 +106,6 @@ def log_if_global(g, g_list, message):
         logger.info(message)  # uses global `logger` instance defined in __main__ script section
 
 
-def composite_colors(r0, g0, b0, a0, r1, g1, b1, a1):
-    """composite a second colour over a first with given alpha values and return the
-    result"""
-    a0 /= 255
-    a1 /= 255
-    a = a0 + a1 - a0 * a1
-    r = (a1 * r1 + (1 - a1) * a0 * r0) / a
-    g = (a1 * g1 + (1 - a1) * a0 * g0) / a
-    b = (a1 * b1 + (1 - a1) * a0 * b0) / a
-    return [int(round(x)) for x in (r, g, b, 255 * a)]
-
-
 @inmain_decorator()
 def error_dialog(message):
     QtWidgets.QMessageBox.warning(app.ui, 'runmanager', message)
@@ -148,607 +140,6 @@ def scroll_view_to_row_if_current(view, item):
     if index.row() == current_row:
         view.scrollTo(index)
         horizontal_scrollbar.setValue(existing_horizontal_position)
-
-
-class FingerTabBarWidget(QtWidgets.QTabBar):
-
-    """A TabBar with the tabs on the left and the text horizontal. Credit to
-    @LegoStormtroopr, https://gist.github.com/LegoStormtroopr/5075267. We will
-    promote the TabBar from the ui file to one of these."""
-
-    def __init__(self, parent=None, minwidth=180, minheight=30, **kwargs):
-        QtWidgets.QTabBar.__init__(self, parent, **kwargs)
-        self.minwidth = minwidth
-        self.minheight = minheight
-        self.iconPosition = kwargs.pop('iconPosition', QtWidgets.QTabWidget.West)
-        self._movable = None
-        self.tab_movable = {}
-        self.paint_clip = None
-
-    def setMovable(self, movable, index=None):
-        """Set tabs movable on an individual basis, or set for all tabs if no
-        index specified"""
-        if index is None:
-            self._movable = movable
-            self.tab_movable = {}
-            QtWidgets.QTabBar.setMovable(self, movable)
-        else:
-            self.tab_movable[int(index)] = bool(movable)
-
-    def isMovable(self, index=None):
-        if index is None:
-            if self._movable is None:
-                self._movable = QtWidgets.QTabBar.isMovable(self)
-            return self._movable
-        return self.tab_movable.get(index, self._movable)
-
-    def indexAtPos(self, point):
-        for index in range(self.count()):
-            if self.tabRect(index).contains(point):
-                return index
-
-    def mouseEventIndex(self, event):
-        if QT_ENV == 'PyQt5':
-            return self.indexAtPos(event.pos())
-        else:
-            # Qt6 position returns QPointF instead of QPoint
-            return self.indexAtPos(event.position().toPoint())
-
-    def mousePressEvent(self, event):
-        index = self.mouseEventIndex(event)
-        if not self.tab_movable.get(index, self.isMovable()):
-            QtWidgets.QTabBar.setMovable(self, False)  # disable dragging until they release the mouse
-        return QtWidgets.QTabBar.mousePressEvent(self, event)
-
-    def mouseReleaseEvent(self, event):
-        if self.isMovable():
-            # Restore this in case it was temporarily disabled by mousePressEvent
-            QtWidgets.QTabBar.setMovable(self, True)
-        return QtWidgets.QTabBar.mouseReleaseEvent(self, event)
-
-    def tabLayoutChange(self):
-        total_height = 0
-        for index in range(self.count()):
-            tabRect = self.tabRect(index)
-            total_height += tabRect.height()
-        if total_height > self.parent().height():
-            # Don't paint over the top of the scroll buttons:
-            scroll_buttons_area_height = 2*max(self.style().pixelMetric(QtWidgets.QStyle.PM_TabBarScrollButtonWidth),
-                                               self.style().pixelMetric(QtWidgets.QStyle.PM_LayoutHorizontalSpacing))
-            self.paint_clip = self.width(), self.parent().height() - scroll_buttons_area_height
-        else:
-            self.paint_clip = None
-
-    def paintEvent(self, event):
-        painter = QtWidgets.QStylePainter(self)
-        if self.paint_clip is not None:
-            painter.setClipRect(0, 0, *self.paint_clip)
-
-        option = QtWidgets.QStyleOptionTab()
-        for index in range(self.count()):
-            tabRect = self.tabRect(index)
-            self.initStyleOption(option, index)
-            painter.drawControl(QtWidgets.QStyle.CE_TabBarTabShape, option)
-            if not self.tabIcon(index).isNull():
-                icon = self.tabIcon(index).pixmap(self.iconSize())
-                alignment = QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter
-                tabRect.moveLeft(10)
-                painter.drawItemPixmap(tabRect, alignment, icon)
-                tabRect.moveLeft(self.iconSize().width() + 15)
-            else:
-                tabRect.moveLeft(10)
-            painter.drawText(tabRect, QtCore.Qt.AlignVCenter, self.tabText(index))
-        if self.paint_clip is not None:
-            x_clip, y_clip = self.paint_clip
-            painter.setClipping(False)
-            palette = self.palette()
-            mid_color = palette.color(QtGui.QPalette.Mid)
-            painter.setPen(mid_color)
-            painter.drawLine(0, y_clip, x_clip, y_clip)
-        painter.end()
-
-
-    def tabSizeHint(self, index):
-        fontmetrics = QtGui.QFontMetrics(self.font())
-        text_size = fontmetrics.size(QtCore.Qt.TextSingleLine, self.tabText(index))
-        text_width = text_size.width()
-        text_height = text_size.height()
-        height = text_height + 15
-        height = max(self.minheight, height)
-        width = text_width + 15
-
-        button = self.tabButton(index, QtWidgets.QTabBar.RightSide)
-        if button is not None:
-            height = max(height, button.height() + 7)
-            # Same amount of space around the button horizontally as it has vertically:
-            width += button.width() + height - button.height()
-        width = max(self.minwidth, width)
-        return QtCore.QSize(width, height)
-
-    def setTabButton(self, index, geometry, button):
-        if not isinstance(button, TabToolButton):
-            raise TypeError('Not a TabToolButton, won\'t paint correctly. Use a TabToolButton')
-        result = QtWidgets.QTabBar.setTabButton(self, index, geometry, button)
-        button.move(*button.get_correct_position())
-        return result
-
-
-class TabToolButton(QtWidgets.QToolButton):
-    def __init__(self, *args, **kwargs):
-        QtWidgets.QToolButton.__init__(self, *args, **kwargs)
-        self.setFocusPolicy(QtCore.Qt.NoFocus)
-
-    def paintEvent(self, event):
-        painter = QtWidgets.QStylePainter(self)
-        paint_clip = self.parent().paint_clip
-        if paint_clip is not None:
-            point = QtCore.QPoint(*paint_clip)
-            global_point = self.parent().mapToGlobal(point)
-            local_point = self.mapFromGlobal(global_point)
-            painter.setClipRect(0, 0, local_point.x(), local_point.y())
-        option = QtWidgets.QStyleOptionToolButton()
-        self.initStyleOption(option)
-        painter.drawComplexControl(QtWidgets.QStyle.CC_ToolButton, option)
-
-    def get_correct_position(self):
-        parent = self.parent()
-        for index in range(parent.count()):
-            if parent.tabButton(index, QtWidgets.QTabBar.RightSide) is self:
-                break
-        else:
-            raise LookupError('Tab not found')
-        tabRect = parent.tabRect(index)
-        tab_x, tab_y, tab_width, tab_height = tabRect.x(), tabRect.y(), tabRect.width(), tabRect.height()
-        size = self.sizeHint()
-        width = size.width()
-        height = size.height()
-        padding = int((tab_height - height) / 2)
-        correct_x = tab_x + tab_width - width - padding
-        correct_y = tab_y + padding
-        return correct_x, correct_y
-
-    def moveEvent(self, event):
-        try:
-            correct_x, correct_y = self.get_correct_position()
-        except LookupError:
-            return # Things aren't initialised yet
-        if self.x() != correct_x or self.y() != correct_y:
-            # Move back! I shall not be moved!
-            self.move(correct_x, correct_y)
-        return QtWidgets.QToolButton.moveEvent(self, event)
-
-
-class FingerTabWidget(QtWidgets.QTabWidget):
-
-    """A QTabWidget equivalent which uses our FingerTabBarWidget"""
-
-    def __init__(self, parent, *args):
-        QtWidgets.QTabWidget.__init__(self, parent, *args)
-        self.setTabBar(FingerTabBarWidget(self))
-
-    def addTab(self, *args, **kwargs):
-        closeable = kwargs.pop('closable', False)
-        index = QtWidgets.QTabWidget.addTab(self, *args, **kwargs)
-        self.setTabClosable(index, closeable)
-        return index
-
-    def setTabClosable(self, index, closable):
-        right_button = self.tabBar().tabButton(index, QtWidgets.QTabBar.RightSide)
-        if closable:
-            if not right_button:
-                # Make one:
-                close_button = TabToolButton(self.parent())
-                close_button.setIcon(QtGui.QIcon(':/qtutils/fugue/cross'))
-                self.tabBar().setTabButton(index, QtWidgets.QTabBar.RightSide, close_button)
-                close_button.clicked.connect(lambda: self._on_close_button_clicked(close_button))
-        else:
-            if right_button:
-                # Get rid of it:
-                self.tabBar().setTabButton(index, QtWidgets.QTabBar.RightSide, None)
-
-    def _on_close_button_clicked(self, button):
-        for index in range(self.tabBar().count()):
-            if self.tabBar().tabButton(index, QtWidgets.QTabBar.RightSide) is button:
-                self.tabCloseRequested.emit(index)
-                break
-
-
-class ItemView(object):
-    """Mixin for QTableView and QTreeView that emits a custom signal leftClicked(index)
-    after a left click on a valid index, and doubleLeftClicked(index) (in addition) on
-    double click. Also has modified tab and arrow key behaviour and custom selection
-    highlighting."""
-    leftClicked = Signal(QtCore.QModelIndex)
-    doubleLeftClicked = Signal(QtCore.QModelIndex)
-
-
-    def __init__(self, *args):
-        super(ItemView, self).__init__(*args)
-        self._pressed_index = None
-        self._double_click = False
-        self.setAutoScroll(False)
-        palette = self.palette()
-        for group in [QtGui.QPalette.Active, QtGui.QPalette.Inactive]:
-            palette.setColor(
-                group,
-                QtGui.QPalette.Highlight,
-                QtGui.QColor(RunmanagerColors().COLOR_HIGHLIGHT)
-            )
-            palette.setColor(
-                group,
-                QtGui.QPalette.HighlightedText,
-                palette.color(QtGui.QPalette.WindowText)
-            )
-        self.setPalette(palette)
-
-    def mouseEventIndex(self, event):
-        if QT_ENV == 'PyQt5':
-            return self.indexAt(event.pos())
-        else:
-            # Qt6 returns QPointF instead of QPoint
-            return self.indexAt(event.position().toPoint())
-
-    def mousePressEvent(self, event):
-        result = super(ItemView, self).mousePressEvent(event)
-        index = self.mouseEventIndex(event)
-        if event.button() == QtCore.Qt.LeftButton and index.isValid():
-            self._pressed_index = self.mouseEventIndex(event)
-        return result
-
-    def leaveEvent(self, event):
-        result = super(ItemView, self).leaveEvent(event)
-        self._pressed_index = None
-        self._double_click = False
-        return result
-
-    def mouseDoubleClickEvent(self, event):
-        # Ensure our left click event occurs regardless of whether it is the
-        # second click in a double click or not
-        result = super(ItemView, self).mouseDoubleClickEvent(event)
-        index = self.mouseEventIndex(event)
-        if event.button() == QtCore.Qt.LeftButton and index.isValid():
-            self._pressed_index = self.mouseEventIndex(event)
-            self._double_click = True
-        return result
-
-    def mouseReleaseEvent(self, event):
-        result = super(ItemView, self).mouseReleaseEvent(event)
-        index = self.mouseEventIndex(event)
-        if event.button() == QtCore.Qt.LeftButton and index.isValid() and index == self._pressed_index:
-            self.leftClicked.emit(index)
-            if self._double_click:
-                self.doubleLeftClicked.emit(index)
-        self._pressed_index = None
-        self._double_click = False
-        return result
-
-    def keyPressEvent(self, event):
-        if event.key() in [QtCore.Qt.Key_Space, QtCore.Qt.Key_Enter, QtCore.Qt.Key_Return]:
-            item = self.model().itemFromIndex(self.currentIndex())
-            if item.isEditable():
-                # Space/enter edits editable items:
-                self.edit(self.currentIndex())
-            else:
-                # Space/enter on non-editable items simulates a left click:
-                self.leftClicked.emit(self.currentIndex())
-        return super(ItemView, self).keyPressEvent(event)
-
-    def moveCursor(self, cursor_action, keyboard_modifiers):
-        current_index = self.currentIndex()
-        current_row, current_column = current_index.row(), current_index.column()
-        if cursor_action == QtWidgets.QAbstractItemView.MoveUp:
-            return current_index.sibling(current_row - 1, current_column)
-        elif cursor_action == QtWidgets.QAbstractItemView.MoveDown:
-            return current_index.sibling(current_row + 1, current_column)
-        elif cursor_action == QtWidgets.QAbstractItemView.MoveLeft:
-            return current_index.sibling(current_row, current_column - 1)
-        elif cursor_action == QtWidgets.QAbstractItemView.MoveRight:
-            return current_index.sibling(current_row, current_column + 1)
-        elif cursor_action == QtWidgets.QAbstractItemView.MovePrevious:
-            return current_index.sibling(current_row, current_column - 1)
-        elif cursor_action == QtWidgets.QAbstractItemView.MoveNext:
-            return current_index.sibling(current_row, current_column + 1)
-        else:
-            return super(ItemView, self).moveCursor(cursor_action, keyboard_modifiers)
-
-
-class TreeView(ItemView, QtWidgets.QTreeView):
-    """Treeview version of our customised ItemView"""
-    def __init__(self, parent=None):
-        super(TreeView, self).__init__(parent)
-        # Set columns to their minimum size, disabling resizing. Caller may still
-        # configure a specific section to stretch:
-        self.header().setSectionResizeMode(
-            QtWidgets.QHeaderView.ResizeToContents
-        )
-        self.setItemDelegate(ItemDelegate(self))
-
-
-class TableView(ItemView, QtWidgets.QTableView):
-    """TableView version of our customised ItemView"""
-    def __init__(self, parent=None):
-        super(TableView, self).__init__(parent)
-        # Set rows and columns to the minimum size, disabling interactive resizing.
-        # Caller may still configure a specific column to stretch:
-        self.verticalHeader().setSectionResizeMode(
-            QtWidgets.QHeaderView.ResizeToContents
-        )
-        self.horizontalHeader().setSectionResizeMode(
-            QtWidgets.QHeaderView.ResizeToContents
-        )
-        self.horizontalHeader().sectionResized.connect(self.on_column_resized)
-        self.setItemDelegate(ItemDelegate(self))
-        self.verticalHeader().hide()
-        self.setShowGrid(False)
-        self.horizontalHeader().setHighlightSections(False)
-
-    def on_column_resized(self, col):
-        for row in range(self.model().rowCount()):
-            self.resizeRowToContents(row)
-
-
-class AlternatingColorModel(QtGui.QStandardItemModel):
-
-    def __init__(self, view):
-        QtGui.QStandardItemModel.__init__(self)
-        # How much darker in each channel is the alternate base color compared
-        # to the base color?
-        self.view = view
-
-        # A cache, store brushes so we don't have to recalculate them. Is faster.
-        self.bg_brushes = {}
-
-    def get_bgbrush(self, normal_brush, alternate, selected):
-        """Get cell colour as a function of its ordinary colour, whether it is on an odd
-        row, and whether it is selected."""
-        normal_rgb = normal_brush.color().getRgb() if normal_brush is not None else None
-        try:
-            return self.bg_brushes[normal_rgb, alternate, selected]
-        except KeyError:
-            pass
-        # Get the colour of the cell with alternate row shading:
-        normal_color = self.view.palette().color(
-            QtGui.QPalette.ColorGroup.Disabled,
-            QtGui.QPalette.ColorRole.Base
-        )
-        alternate_color = self.view.palette().color(
-            QtGui.QPalette.ColorGroup.Disabled,
-            QtGui.QPalette.ColorRole.AlternateBase
-        )
-        if normal_rgb is None:
-            # No colour has been set. Use palette colours:
-            if alternate:
-                bg_color = alternate_color
-            else:
-                bg_color = normal_color
-        else:
-            bg_color = normal_brush.color()
-            if alternate:
-                # Modify alternate rows:
-                r, g, b, a = normal_rgb
-                nr, ng, nb, na = normal_color.getRgb()
-                ar, ag, ab, aa = alternate_color.getRgb()
-                alt_r = min(max(r + ar - nr, 0), 255)
-                alt_g = min(max(g + ag - ng, 0), 255)
-                alt_b = min(max(b + ab - nb, 0), 255)
-                alt_a = min(max(a + aa - na, 0), 255)
-                bg_color = QtGui.QColor(alt_r, alt_g, alt_b, alt_a)
-
-        # If parent is a TableView, we handle selection highlighting as part of the
-        # background colours:
-        if selected and isinstance(self.view, QtWidgets.QTableView):
-            # Overlay highlight colour:
-            r_s, g_s, b_s, a_s = QtGui.QColor(RunmanagerColors().COLOR_HIGHLIGHT).getRgb()
-            r_0, g_0, b_0, a_0 = bg_color.getRgb()
-            rgb = composite_colors(r_0, g_0, b_0, a_0, r_s, g_s, b_s, a_s)
-            bg_color = QtGui.QColor(*rgb)
-
-        brush = QtGui.QBrush(bg_color)
-        self.bg_brushes[normal_rgb, alternate, selected] = brush
-        return brush
-
-    def data(self, index, role):
-        """When background color data is being requested, returns modified colours for
-        every second row, according to the palette of the view. This has the effect of
-        making the alternate colours visible even when custom colors have been set - the
-        same shading will be applied to the custom colours. Only really looks sensible
-        when the normal and alternate colors are similar. Also applies selection
-        highlight colour (using RunmanagerColors().COLOR_HIGHLIGHT), similarly with alternate-row
-        shading, for the case of a QTableView."""
-        if role == QtCore.Qt.BackgroundRole:
-            normal_brush = QtGui.QStandardItemModel.data(self, index, QtCore.Qt.BackgroundRole)
-            selected = index in self.view.selectedIndexes()
-            alternate = index.row() % 2
-            return self.get_bgbrush(normal_brush, alternate, selected)
-        return QtGui.QStandardItemModel.data(self, index, role)
-
-
-class Editor(QtWidgets.QTextEdit):
-    """Popup editor with word wrapping and automatic resizing."""
-
-    def __init__(self, parent):
-        QtWidgets.QTextEdit.__init__(self, parent)
-        self.setWordWrapMode(QtGui.QTextOption.WordWrap)
-        self.setAcceptRichText(False)
-        self.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
-        self.setHorizontalScrollBarPolicy(QtCore.Qt.ScrollBarAlwaysOff)
-        self.textChanged.connect(self.update_size)
-        self.initial_height = None
-        palette = self.palette()
-        for group in [QtGui.QPalette.Active, QtGui.QPalette.Inactive]:
-            palette.setColor(
-                group,
-                QtGui.QPalette.Highlight,
-                QtGui.QColor(RunmanagerColors().COLOR_HIGHLIGHT)
-            )
-            palette.setColor(
-                group,
-                QtGui.QPalette.HighlightedText,
-                palette.color(QtGui.QPalette.WindowText)
-            )
-        self.setPalette(palette)
-
-    def update_size(self):
-        if self.initial_height is not None:
-            # Temporarily shrink back to the initial height, just so that the document
-            # size below returns the preferred size rather than the current size.
-            # QTextDocument doesn't have a sizeHint of minimumSizeHint method, so this
-            # is the best we can do to get its minimum size.
-            self.setFixedHeight(self.initial_height)
-        preferred_height = self.document().size().toSize().height()
-        # Do not shrink smaller than the initial height:
-        if self.initial_height is not None and preferred_height >= self.initial_height:
-            self.setFixedHeight(preferred_height)
-
-    def resizeEvent(self, event):
-        result = QtWidgets.QTextEdit.resizeEvent(self, event)
-        # Record the initial height after it is first set:
-        if self.initial_height is None:
-            self.initial_height = self.height()
-        return result
-        
-
-
-class ItemDelegate(QtWidgets.QStyledItemDelegate):
-
-    """An item delegate with a larger row height and column width, faint grey vertical
-    lines between columns, and a custom editor for handling multi-line data"""
-    MIN_ROW_HEIGHT = 22
-    EXTRA_ROW_HEIGHT = 6
-    EXTRA_COL_WIDTH = 20
-
-    def __init__(self, *args, **kwargs):
-        QtWidgets.QStyledItemDelegate.__init__(self, *args, **kwargs)
-        self._pen = QtGui.QPen()
-        self._pen.setWidth(1)
-        self._pen.setColor(QtGui.QColor.fromRgb(128, 128, 128, 64))
-
-    def sizeHint(self, *args):
-        size = QtWidgets.QStyledItemDelegate.sizeHint(self, *args)
-        if size.height() <= self.MIN_ROW_HEIGHT:
-            height = self.MIN_ROW_HEIGHT
-        else:
-            # Esnure cells with multiple lines of text still have some padding:
-            height = size.height() + self.EXTRA_ROW_HEIGHT
-        return QtCore.QSize(size.width() + self.EXTRA_COL_WIDTH, height)
-
-    def paint(self, painter, option, index):
-        if isinstance(self.parent(), QtWidgets.QTableView):
-            # Disable rendering of selection highlight for TableViews, they handle
-            # it themselves with the background colour data:
-            option.state &= ~(QtWidgets.QStyle.State_Selected)
-        QtWidgets.QStyledItemDelegate.paint(self, painter, option, index)
-        if index.column() > 0:
-            painter.setPen(self._pen)
-            painter.drawLine(option.rect.topLeft(), option.rect.bottomLeft())
-
-    def eventFilter(self, obj, event):
-        """Filter events before they get to the editor, so that editing is ended when
-        the user presses tab, shift-tab or enter (which otherwise would not end editing
-        in a QTextEdit)."""
-        if event.type() == QtCore.QEvent.KeyPress:
-            if event.key() in [QtCore.Qt.Key_Enter, QtCore.Qt.Key_Return]:
-                # Allow shift-enter
-                if not event.modifiers() & QtCore.Qt.ShiftModifier:
-                    self.commitData.emit(obj)
-                    self.closeEditor.emit(obj)
-                    return True
-            elif event.key() == QtCore.Qt.Key_Tab:
-                self.commitData.emit(obj)
-                self.closeEditor.emit(obj, QtWidgets.QStyledItemDelegate.EditNextItem)
-                return True
-            elif event.key() == QtCore.Qt.Key_Backtab:
-                self.commitData.emit(obj)
-                self.closeEditor.emit(obj, QtWidgets.QStyledItemDelegate.EditPreviousItem)
-                return True
-        return QtWidgets.QStyledItemDelegate.eventFilter(self, obj, event)
-
-    def createEditor(self, parent, option, index):
-        return Editor(parent)
-
-    def setEditorData(self, editor, index):
-        editor.setPlainText(index.data())
-        font = index.data(QtCore.Qt.FontRole)
-        default_font = qapplication.font(self.parent())
-        if font is None:
-            font = default_font
-        font.setPointSize(default_font.pointSize())
-        editor.setFont(font)
-        font_height = QtGui.QFontMetrics(font).height()
-        padding = (self.MIN_ROW_HEIGHT - font_height) / 2 - 1
-        editor.document().setDocumentMargin(padding)
-        editor.selectAll()
-        
-    def setModelData(self, editor, model, index):
-        model.setData(index, editor.toPlainText())
-
-class RunmanagerColors(object):
-    """Singleton class that globally defines various colors for the globals view
-    
-    Colors are saved to class attributes as hex strings.
-    Available colors are:
-
-    :ivar COLOR_HIGHLIGHT: Item selection highlight color, fixed to semitransparent blue
-    :ivar COLOR_ERROR: Item has runtime error color
-    :ivar COLOR_OK: Item is normal
-    :ivar COLOR_BOOL_ON: Item is bool = True
-    :ivar COLOR_BOOL_OFF: Item is bool = False
-    """
-    _instance = None
-
-    def __new__(cls, *args, **kwargs):
-        if cls._instance is None:
-            cls._instance = super().__new__(cls)
-
-        return cls._instance
-    
-    def __init__(self):
-
-        if not hasattr(self, '_initialized'):
-            self.logger = logging.getLogger('runmanager')
-
-            # init colors
-            self.update_colors_from_scheme()
-
-            # common color definitions
-            self.COLOR_HIGHLIGHT = "#40308CC6"  # semitransparent blue
-
-            self._initialized = True
-    
-    def check_if_light(self):
-        """Helper method that return current light/dark theme state
-        
-        If using PyQt5, always returns True.
-        Otherwise, state is taken from `QGuiApplication` which follows OS.
-        """
-        
-        # pyqt5 defaults to light and doesn't have styleHints.colorScheme()
-        if QT_ENV.lower() == 'pyqt5':
-            return True
-        
-        style_hints = QtGui.QGuiApplication.styleHints()
-        if style_hints.colorScheme() == QtCore.Qt.ColorScheme.Dark:
-            return False
-        else:
-            return True
-        
-    def update_colors_from_scheme(self):
-        """Method updates colors depending on current color scheme"""
-
-        if self.check_if_light():
-            self.logger.info('Setting custom light theme colors')
-            # use light mode colors for GroupTabs
-            self.COLOR_ERROR = '#F79494'  # light red
-            self.COLOR_OK = '#A5F7C6'  # light green
-            self.COLOR_BOOL_ON = '#63F731'  # bright green
-            self.COLOR_BOOL_OFF = '#608060'  # dark green
-        else:
-            self.logger.info('Setting custom dark theme colors')
-            # use dark mode colors for GroupTabs
-            self.COLOR_ERROR = "#BC0000"  # red
-            self.COLOR_OK = "#2F4C00"  # green
-            self.COLOR_BOOL_ON = "#29A300"  # bright green
-            self.COLOR_BOOL_OFF = "#003900"  # dark green
 
 
 class GroupTab(object):
@@ -864,7 +255,7 @@ class GroupTab(object):
             self.tabWidget.setTabIcon(index, icon)
 
     def populate_model(self):
-        globals = runmanager.get_globals({self.group_name: self.globals_file})[self.group_name]
+        globals = group_manager.get_globals({self.group_name: self.globals_file})[self.group_name]
         for name, (value, units, expansion) in globals.items():
             row = self.make_global_row(name, value, units, expansion)
             self.globals_model.appendRow(row)
@@ -1148,7 +539,7 @@ class GroupTab(object):
         item = self.get_global_item_by_name(global_name, self.GLOBALS_COL_NAME,
                                             previous_name=self.GLOBALS_DUMMY_ROW_TEXT)
         try:
-            runmanager.new_global(self.globals_file, self.group_name, global_name)
+            group_manager.new_global(self.globals_file, self.group_name, global_name)
         except Exception as e:
             error_dialog(str(e))
         else:
@@ -1175,7 +566,7 @@ class GroupTab(object):
         item = self.get_global_item_by_name(new_global_name, self.GLOBALS_COL_NAME,
                                             previous_name=previous_global_name)
         try:
-            runmanager.rename_global(self.globals_file, self.group_name, previous_global_name, new_global_name)
+            group_manager.rename_global(self.globals_file, self.group_name, previous_global_name, new_global_name)
         except Exception as e:
             error_dialog(str(e))
             # Set the item text back to the old name, since the rename failed:
@@ -1219,7 +610,7 @@ class GroupTab(object):
 
     def complete_change_global_value(self, global_name, previous_value, new_value, item, previous_background, previous_icon, interactive=True):
         try:
-            runmanager.set_value(self.globals_file, self.group_name, global_name, new_value)
+            group_manager.set_value(self.globals_file, self.group_name, global_name, new_value)
         except Exception as e:
             if interactive:
                 error_dialog(str(e))
@@ -1255,7 +646,7 @@ class GroupTab(object):
                     (self.globals_file, self.group_name, global_name, previous_units, new_units))
         item = self.get_global_item_by_name(global_name, self.GLOBALS_COL_UNITS)
         try:
-            runmanager.set_units(self.globals_file, self.group_name, global_name, new_units)
+            group_manager.set_units(self.globals_file, self.group_name, global_name, new_units)
         except Exception as e:
             error_dialog(str(e))
             # Set the item text back to the old units, since the change failed:
@@ -1272,7 +663,7 @@ class GroupTab(object):
                     (self.globals_file, self.group_name, global_name, previous_expansion, new_expansion))
         item = self.get_global_item_by_name(global_name, self.GLOBALS_COL_EXPANSION)
         try:
-            runmanager.set_expansion(self.globals_file, self.group_name, global_name, new_expansion)
+            group_manager.set_expansion(self.globals_file, self.group_name, global_name, new_expansion)
         except Exception as e:
             error_dialog(str(e))
             # Set the item text back to the old units, since the change failed:
@@ -1349,7 +740,7 @@ class GroupTab(object):
         if confirm:
             if not question_dialog("Delete the global '%s'?" % global_name):
                 return
-        runmanager.delete_global(self.globals_file, self.group_name, global_name)
+        group_manager.delete_global(self.globals_file, self.group_name, global_name)
         # Find the entry for this global in self.globals_model and remove it:
         name_item = self.get_global_item_by_name(global_name, self.GLOBALS_COL_NAME)
         self.globals_model.removeRow(name_item.row())
@@ -2323,7 +1714,7 @@ class RunManager(object):
         # Save the containing folder for use next time we open the dialog box:
         self.last_opened_globals_folder = os.path.dirname(globals_file)
         # Create the new file and open it:
-        runmanager.new_globals_file(globals_file)
+        group_manager.new_globals_file(globals_file)
         self.open_globals_file(globals_file)
 
     def on_diff_globals_file_clicked(self):
@@ -2348,7 +1739,7 @@ class RunManager(object):
             return
 
         # Get file's globals groups
-        other_groups = runmanager.get_all_groups(globals_file)
+        other_groups = group_manager.get_all_groups(globals_file)
 
         # Display the output tab so the user can see the output:
         self.ui.tabWidget.setCurrentWidget(self.ui.tab_output)
@@ -2590,7 +1981,7 @@ class RunManager(object):
         current_labscript_file = self.ui.lineEdit_labscript_file.text()
         if not current_labscript_file:
             return ''
-        _, default_output_folder, _ = runmanager.new_sequence_details(
+        _, default_output_folder, _ = compiler.new_sequence_details(
             current_labscript_file,
             config=self.exp_config,
             increment_sequence_index=False,
@@ -2862,7 +2253,7 @@ class RunManager(object):
             return
 
         # Get the groups:
-        groups = runmanager.get_grouplist(globals_file)
+        groups = group_manager.get_grouplist(globals_file)
         # Add the parent row:
         file_name_item = QtGui.QStandardItem(globals_file)
         file_name_item.setEditable(False)
@@ -2993,7 +2384,7 @@ class RunManager(object):
         if delete_source_group and source_globals_file == dest_globals_file:
             return
         try:
-            dest_group_name = runmanager.copy_group(source_globals_file, source_group_name, dest_globals_file, delete_source_group)
+            dest_group_name = group_manager.copy_group(source_globals_file, source_group_name, dest_globals_file, delete_source_group)
         except Exception as e:
             error_dialog(str(e))
         else:
@@ -3031,7 +2422,7 @@ class RunManager(object):
         item = self.get_group_item_by_name(globals_file, group_name, self.GROUPS_COL_NAME,
                                            previous_name=self.GROUPS_DUMMY_ROW_TEXT)
         try:
-            runmanager.new_group(globals_file, group_name)
+            group_manager.new_group(globals_file, group_name)
         except Exception as e:
             error_dialog(str(e))
         else:
@@ -3076,7 +2467,7 @@ class RunManager(object):
         item = self.get_group_item_by_name(globals_file, new_group_name, self.GROUPS_COL_NAME,
                                            previous_name=previous_group_name)
         try:
-            runmanager.rename_group(globals_file, previous_group_name, new_group_name)
+            group_manager.rename_group(globals_file, previous_group_name, new_group_name)
         except Exception as e:
             error_dialog(str(e))
             # Set the item text back to the old name, since the rename failed:
@@ -3109,7 +2500,7 @@ class RunManager(object):
         group_tab = self.currently_open_groups.get((globals_file, group_name))
         if group_tab is not None:
             self.close_group(globals_file, group_name)
-        runmanager.delete_group(globals_file, group_name)
+        group_manager.delete_group(globals_file, group_name)
         # Find the entry for this group in self.groups_model and remove it:
         name_item = self.get_group_item_by_name(globals_file, group_name, self.GROUPS_COL_NAME)
         name_item.parent().removeRow(name_item.row())
@@ -3399,15 +2790,15 @@ class RunManager(object):
                 continue
 
     def parse_globals(self, active_groups, raise_exceptions=True, expand_globals=True, expansion_order = None, return_dimensions = False):
-        sequence_globals = runmanager.get_globals(active_groups)
+        sequence_globals = group_manager.get_globals(active_groups)
         #logger.info('got sequence globals')
-        evaled_globals, global_hierarchy, expansions = runmanager.evaluate_globals(sequence_globals, raise_exceptions)
+        evaled_globals, global_hierarchy, expansions = group_manager.evaluate_globals(sequence_globals, raise_exceptions)
         #logger.info('evaluated sequence globals')
         if expand_globals:
             if return_dimensions:
-                shots, dimensions = runmanager.expand_globals(sequence_globals, evaled_globals, expansion_order, return_dimensions=return_dimensions)
+                shots, dimensions = compiler.expand_globals(sequence_globals, evaled_globals, expansion_order, return_dimensions=return_dimensions)
             else:
-                shots = runmanager.expand_globals(sequence_globals, evaled_globals, expansion_order)
+                shots = compiler.expand_globals(sequence_globals, evaled_globals, expansion_order)
         else:
             shots = []
             dimensions = {}
@@ -3443,7 +2834,7 @@ class RunManager(object):
                     # Let ExpansionErrors through through, as they occur
                     # when the user has changed the value without changing
                     # the expansion type:
-                    if isinstance(value, runmanager.ExpansionError):
+                    if isinstance(value, ExpansionError):
                         continue
                     return False
         # Did the guessed expansion type for any of the globals change?
@@ -3474,8 +2865,8 @@ class RunManager(object):
                     else:
                         previous_value = 0
 
-                new_guess = runmanager.guess_expansion_type(new_value)
-                previous_guess = runmanager.guess_expansion_type(previous_value)
+                new_guess = group_manager.guess_expansion_type(new_value)
+                previous_guess = group_manager.guess_expansion_type(previous_value)
 
                 if new_guess == 'outer':
                     expansion_types[global_name] = {'previous_guess': previous_guess,
@@ -3485,7 +2876,7 @@ class RunManager(object):
                                                     }
                 elif new_guess != previous_guess:
                     filename = active_groups[group_name]
-                    runmanager.set_expansion(filename, group_name, global_name, new_guess)
+                    group_manager.set_expansion(filename, group_name, global_name, new_guess)
                     expansions[global_name] = new_guess
                     expansion_types_changed = True
 
@@ -3534,7 +2925,7 @@ class RunManager(object):
             # we have a global that does not depend on anything that has an
             # expansion type of 'outer'
             if (not global_depends_on_global_with_outer_product(global_name, global_hierarchy, expansions)
-                    and not isinstance(expansion_types[global_name]['value'], runmanager.ExpansionError)):
+                    and not isinstance(expansion_types[global_name]['value'], ExpansionError)):
                 current_dependencies = find_dependencies(global_name, global_hierarchy, expansion_types)
 
                 # if this global has other globals that use it, then add them
@@ -3548,7 +2939,7 @@ class RunManager(object):
         for global_name in sorted(self.previous_expansion_types):
             if (not global_depends_on_global_with_outer_product(
                 global_name, self.previous_global_hierarchy, self.previous_expansions)
-                    and not isinstance(self.previous_expansion_types[global_name]['value'], runmanager.ExpansionError)):
+                    and not isinstance(self.previous_expansion_types[global_name]['value'], ExpansionError)):
                 old_dependencies = find_dependencies(global_name, self.previous_global_hierarchy, self.previous_expansion_types)
                 # if this global has other globals that use it, then add them
                 # all to a zip group with the name of this global
@@ -3562,7 +2953,7 @@ class RunManager(object):
         for global_name, guesses in expansion_types.items():
             if guesses['new_guess'] != guesses['previous_guess']:
                 filename = active_groups[guesses['group_name']]
-                runmanager.set_expansion(
+                group_manager.set_expansion(
                     filename, str(guesses['group_name']), str(global_name), str(guesses['new_guess']))
                 expansions[global_name] = guesses['new_guess']
                 expansion_types_changed = True
@@ -3576,7 +2967,7 @@ class RunManager(object):
                         iter(evaled_globals[group_name][global_name])
                     except Exception:
                         filename = active_groups[group_name]
-                        runmanager.set_expansion(filename, group_name, global_name, '')
+                        group_manager.set_expansion(filename, group_name, global_name, '')
                         expansion_types_changed = True
 
         self.previous_evaled_globals = evaled_globals
@@ -3587,7 +2978,7 @@ class RunManager(object):
         return expansion_types_changed
 
     def make_h5_files(self, labscript_file, output_folder, sequence_globals, shots, shuffle):
-        sequence_attrs, default_output_dir, filename_prefix = runmanager.new_sequence_details(
+        sequence_attrs, default_output_dir, filename_prefix = compiler.new_sequence_details(
             labscript_file, config=self.exp_config, increment_sequence_index=True
         )
         if output_folder == self.previous_default_output_folder:
@@ -3597,7 +2988,7 @@ class RunManager(object):
             # from the UI may be out of date since we only update it once a second.
             output_folder = default_output_dir
         self.check_output_folder_update()
-        run_files = runmanager.make_run_files(
+        run_files = compiler.make_run_files(
             output_folder,
             sequence_globals,
             shots,
@@ -3669,14 +3060,14 @@ class RemoteServer(ZMQServer):
 
     def handle_get_globals(self, raw=False):
         active_groups = inmain(app.get_active_groups, interactive=False)
-        sequence_globals = runmanager.get_globals(active_groups)
+        sequence_globals = group_manager.get_globals(active_groups)
         all_globals = {}
         if raw:
             for group_globals in sequence_globals.values():
                 values_only = {name: val for name, (val, _, _) in group_globals.items()}
                 all_globals.update(values_only)
         else:
-            evaled_globals, global_hierarchy, expansions = runmanager.evaluate_globals(
+            evaled_globals, global_hierarchy, expansions = group_manager.evaluate_globals(
                 sequence_globals, raise_exceptions=False
             )
             for group_globals in evaled_globals.values():
@@ -3686,7 +3077,7 @@ class RemoteServer(ZMQServer):
     @inmain_decorator()
     def handle_set_globals(self, globals, raw=False):
         active_groups = app.get_active_groups(interactive=False)
-        sequence_globals = runmanager.get_globals(active_groups)
+        sequence_globals = group_manager.get_globals(active_groups)
         try:
             for global_name, new_value in globals.items():
                 # Unless raw=True, convert to str representation for saving to the GUI
@@ -3728,7 +3119,7 @@ class RemoteServer(ZMQServer):
                             ]
                         except KeyError:
                             # Group is not open. Change the global value on disk:
-                            runmanager.set_value(
+                            group_manager.set_value(
                                 globals_file, group_name, global_name, new_value
                             )
                         else:
@@ -3813,9 +3204,9 @@ class RemoteServer(ZMQServer):
             # This will raise an exception if there are multiple active groups of the
             # same name:
             active_groups = inmain(app.get_active_groups, interactive=False)
-            sequence_globals = runmanager.get_globals(active_groups)
+            sequence_globals = group_manager.get_globals(active_groups)
             # This will raise an exception if any of the globals can't be evaluated:
-            runmanager.evaluate_globals(sequence_globals, raise_exceptions=True)
+            group_manager.evaluate_globals(sequence_globals, raise_exceptions=True)
         except Exception:
             return True
         return False
