+++ runmanager/exceptions.py
@@ -1,0 +1,5 @@
+class ExpansionError(Exception):
+
+    """An exception class so that error handling code can tell when a
+    parsing exception was caused by a mismatch with the expansion mode"""
+    pass
